pyproject.toml
[tool.poetry]
name = "discordpy-bot-template-with-commands-extension"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.3.4"
flask = "^1.1.2"
bs4 = "^0.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
